==> 3.Code/2.BalancedRandomForest.ipynb <==
import pandas as pd
from numpy import mean
from sklearn.model_selection import train_test_split
from imblearn.ensemble import BalancedRandomForestClassifier
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import RepeatedStratifiedKFold
from sklearn.metrics import classification_report
from sklearn.ensemble import RandomForestClassifier
import joblib
import time

## Load Features and Targets

FF=pd.read_csv('1. Features_and_targets\df_fan_feature.csv')
FP=pd.read_csv('1. Features_and_targets\df_pump_feature.csv')
FS=pd.read_csv('1. Features_and_targets\df_slider_feature.csv')
FV=pd.read_csv('1. Features_and_targets\df_valve_feature.csv')
TF=pd.read_csv('1. Features_and_targets\df_fan_target.csv')
TP=pd.read_csv('1. Features_and_targets\df_pump_target.csv')
TS=pd.read_csv('1. Features_and_targets\df_slider_target.csv')
TV=pd.read_csv('1. Features_and_targets\df_valve_target.csv')
Fall=pd.read_csv('1. Features_and_targets\df_all_features.csv')
Tall=pd.read_csv('1. Features_and_targets\df_all_targets.csv')
Machine=pd.read_csv('1. Features_and_targets\mach_type.csv')


## Dataset splitting

#Slit Dataset in 3. Test (70%) , Train ( 2%) and Validation set (10%)
def split(F, T):
    X_train , X_test, y_train, y_test  = train_test_split( F, T.values.ravel(), random_state=42)
    X_test, X_val, y_test, y_val = train_test_split(X_test, y_test, test_size=0.33, random_state=42)
    return X_train, y_train, X_test, y_test, X_val, y_val

#For all
X_train_all, y_train_all, X_test_all, y_test_all, X_val_all, y_val_all=split(Fall, Tall)
#For FAN
X_train_fan, y_train_fan, X_test_fan, y_test_fan, X_val_fan, y_val_fan=split(FF, TF)
#For VALVE
X_train_valve, y_train_valve, X_test_valve, y_test_valve, X_val_valve, y_val_valve=split(FV, TV)
#For PUMP
X_train_pump, y_train_pump, X_test_pump, y_test_pump, X_val_pump, y_val_pump=split(FP, TP)
#For SLIDER
X_train_slider, y_train_slider, X_test_slider, y_test_slider, X_val_slider, y_val_slider=split(FS, TS)


## Models

def Balanced_RF(a, b, c, d, e, f): # a, b= Training sets / c,d = testing sets / e, f = validation sets
    start = time.process_time()
    model = BalancedRandomForestClassifier(n_estimators=200)
    #X_train, y_train
    model.fit(a, b)
    #evaluation procedure
    cv = RepeatedStratifiedKFold(n_splits=10, n_repeats=3, random_state=42)
    #model evaluation
    scores= cross_val_score(model, c, d, scoring='roc_auc', cv=cv, n_jobs=-1)
    print(f'Random Forest Model\'s accuracy on test set is {100*model.score(X_test_all, y_test_all):.2f}%')
    print('Mean ROC AUC: %.3f' % mean(scores))
    #first report
    y_pred = model.predict(c)
    print('Report  Test / Test Predict')
    print(classification_report(d, y_pred))
     #second report
    y_pred_val = model.predict(e)
    print('Report  Validation / Test Validation')
    print(classification_report(f, y_pred_val))
    print('Processing time:',time.process_time() - start)
    return model

### ALL
#### 0 = Abnormal / 1 = Normal 

model_all=Balanced_RF(X_train_all, y_train_all, X_test_all, y_test_all, X_val_all, y_val_all)

### FAN

model_fan=Balanced_RF(X_train_fan, y_train_fan, X_test_fan, y_test_fan, X_val_fan, y_val_fan)

### VALVE

model_valve=Balanced_RF(X_train_valve, y_train_valve, X_test_valve, y_test_valve, X_val_valve, y_val_valve)


### PUMP

model_pump=Balanced_RF(X_train_pump, y_train_pump, X_test_pump, y_test_pump, X_val_pump, y_val_pump)



### SLIDER

model_slider=Balanced_RF(X_train_slider, y_train_slider, X_test_slider, y_test_slider, X_val_slider, y_val_slider)

## Machine type pred

X_train_m, y_train_m, X_test_m, y_test_m, X_val_m, y_val_m=split(Fall, Machine)

def normal_RF(a, b, c, d, e, f):
    start = time.process_time()
    model = RandomForestClassifier(n_estimators = 50, criterion ='entropy', warm_start = True, max_features = 'sqrt', oob_score = 'True', random_state=42)
    #X_train, y_train
    model.fit(a, b)
    print(f'Random Forest Model\'s accuracy on training set is {100*model.score(a, b):.2f}%')
    print(f'Random Forest Model\'s accuracy on test set is {100*model.score(c, d):.2f}%')

    #first report
    y_pred = model.predict(c)
    print('Report  Test / Test Predict')
    print(classification_report(d, y_pred))
     #second report
    y_pred_val = model.predict(e)
    print('Report  Validation / Test Validation')
    print(classification_report(f, y_pred_val))
    print('Processing time:',time.process_time() - start)
    return model

1: FAN, 2:PUMP, 3:SLIDER, 4:VALVE

model_machine=normal_RF(X_train_m, y_train_m, X_test_m, y_test_m, X_val_m, y_val_m)

## Model export

joblib.dump(model_all, filename='model_all.sav', compress=3)
joblib.dump(model_fan, filename='model_fan.sav', compress=3)
joblib.dump(model_valve, filename='model_valve.sav', compress=3)
joblib.dump(model_pump, filename='model_pump.sav', compress=3)
joblib.dump(model_slider, filename='model_slider.sav', compress=3)
joblib.dump(model_machine, filename='model_machine.sav', compress=3)
